requirements.txt
instaloader==4.14
requests==2.32.3
python-dotenv==1.0.1
free-proxy==1.1.3
